@@ -1,6 +1,8 @@
 project(test.simfil)
 
-set(CMAKE_CXX_STANDARD 17) # Required for catch2
+set(CMAKE_CXX_STANDARD 20) # Required for catch2
+target_compile_features(Catch2 PRIVATE cxx_std_20)
+target_compile_features(Catch2WithMain PRIVATE cxx_std_20)
 
 add_executable(test.simfil
   common.hpp
@@ -11,7 +13,10 @@ add_executable(test.simfil
   completion.cpp
   complex.cpp
   performance.cpp
-  arena.cpp)
+  arena.cpp
+  error.cpp
+  value.cpp
+  operator.cpp)
 
 target_link_libraries(test.simfil
   PUBLIC
