@@ -1,11 +1,11 @@
 #pragma once
 
-#include "tl/expected.hpp"
+#include <tl/expected.hpp>
 
 // Helper macro for bubbling-up tl::expected errors.
 #define TRY_EXPECTED(res)                                                      \
   do {                                                                         \
-    if (!(res).has_value()) {                                                  \
+    if (!(res).has_value()) [[unlikely]] {                                     \
       return tl::unexpected(std::move((res).error()));                         \
     }                                                                          \
   } while (false)
