@@ -1,6 +1,7 @@
 #include "simfil/simfil.h"
 #include "simfil/model/nodes.h"
 #include "simfil/model/string-pool.h"
+#include "simfil/sourcelocation.h"
 #include "simfil/token.h"
 #include "simfil/operator.h"
 #include "simfil/value.h"
@@ -14,6 +15,7 @@
 #include "fmt/core.h"
 
 #include "expressions.h"
+#include "expression-patterns.h"
 #include "completion.h"
 #include "expected.h"
 
@@ -64,44 +66,6 @@ enum Precedence {
 };
 
 /**
- *
- */
-template <class ...Type>
-static auto expect(const ExprPtr& e, Type... types) -> std::optional<unexpected<Error>>
-{
-    const auto type2str = [](Expr::Type t) {
-        switch (t) {
-        case Expr::Type::FIELD:     return "field"s;
-        case Expr::Type::PATH:      return "path"s;
-        case Expr::Type::SUBEXPR:   return "subexpression"s;
-        case Expr::Type::SUBSCRIPT: return "subscript"s;
-        case Expr::Type::VALUE:     return "value"s;
-        }
-        return "error"s;
-    };
-
-    if (!e)
-        return unexpected<Error>(Error::InvalidExpression, "Expected expression");
-
-    if constexpr (sizeof...(types) >= 1) {
-        Expr::Type list[] = {types...};
-        for (auto i = 0; i < sizeof...(types); ++i) {
-            if (e->type() == list[i])
-                return {};
-        }
-
-        std::string typeNames;
-        for (auto i = 0; i < sizeof...(types); ++i) {
-            if (!typeNames.empty())
-                typeNames += " or ";
-            typeNames += type2str(list[i]);
-        }
-
-        return unexpected<Error>(Error::InvalidExpression, fmt::format("Expected {} got {}", typeNames, type2str(e->type())));
-    }
-}
-
-/**
  * Returns if a word should be parsed as a symbol (string).
  * This is true for all UPPER_CASE words.
  */
@@ -120,14 +84,17 @@ static auto isSymbolWord(std::string_view sv) -> bool
 /**
  * RIIA Helper for calling function at destruction.
  */
+template <class Fun>
 struct scoped {
-    std::function<void()> f;
+    Fun f;
+    bool call = true;
 
-    explicit scoped(std::function<void()> f) : f(std::move(f)) {}
-    scoped(scoped&& s) noexcept : f(std::move(s.f)) { s.f = nullptr; }
+    explicit scoped(Fun f) : f(std::move(f)) {}
+    scoped(scoped&& s) noexcept : f(std::move(s.f)) { s.call = false; }
     scoped(const scoped& s) = delete;
     ~scoped() {
-        try { if (f) { f(); } } catch (...) {}
+        if (call)
+            f();
     }
 };
 
@@ -157,9 +124,9 @@ static auto simplifyOrForward(Environment* env, expected<ExprPtr, Error> expr) -
 
     std::deque<Value> values;
     auto stub = Context(env, Context::Phase::Compilation);
-    (void)(*expr)->eval(stub, Value::undef(), LambdaResultFn([&, n = 0](Context ctx, Value vv) mutable {
+    auto res = (*expr)->eval(stub, Value::undef(), LambdaResultFn([&, n = 0](Context ctx, Value&& vv) mutable {
         n += 1;
-        if ((n <= MultiConstExpr::Limit) && (!vv.isa(ValueType::Undef) || vv.node)) {
+        if ((n <= MultiConstExpr::Limit) && (!vv.isa(ValueType::Undef) || vv.nodePtr())) {
             values.push_back(std::move(vv));
             return Result::Continue;
         }
@@ -167,12 +134,13 @@ static auto simplifyOrForward(Environment* env, expected<ExprPtr, Error> expr) -
         values.clear();
         return Result::Stop;
     }));
+    TRY_EXPECTED(res);
 
     /* Warn about constant results */
     if (!values.empty() && std::ranges::all_of(values.begin(), values.end(), [](const Value& v) {
         return v.isa(ValueType::Null);
     }))
-        env->warn("Expression is alway null"s, (*expr)->toString());
+        env->warn("Expression is always null"s, (*expr)->toString());
 
     if (!values.empty() && values[0].isa(ValueType::Bool) && std::ranges::all_of(values.begin(), values.end(), [&](const Value& v) {
         return v.isBool(values[0].as<ValueType::Bool>());
@@ -478,9 +446,6 @@ class SubscriptParser : public PrefixParselet, public InfixParselet
 
     auto parse(Parser& p, ExprPtr left, Token t) const -> expected<ExprPtr, Error> override
     {
-        if (auto err = expect(left, Expr::Type::PATH, Expr::Type::VALUE, Expr::Type::SUBEXPR, Expr::Type::SUBSCRIPT))
-            return *err;
-
         auto _ = scopedNotInPath(p);
         auto body = p.parseTo(Token::RBRACK);
         if (!body)
@@ -510,23 +475,16 @@ class SubSelectParser : public PrefixParselet, public InfixParselet
         /* Prefix sub-selects are transformed to a right side path expression,
          * with the current node on the left. As "standalone" sub-selects are not useful. */
         auto body = p.parseTo(Token::RBRACE);
-        if (!body)
-            return unexpected<Error>(std::move(body.error()));
-
+        TRY_EXPECTED(body);
         return simplifyOrForward(p.env, std::make_unique<SubExpr>(std::make_unique<FieldExpr>("_"),
                                                                   std::move(*body)));
     }
 
     auto parse(Parser& p, ExprPtr left, Token t) const -> expected<ExprPtr, Error> override
     {
-        if (auto err = expect(left, Expr::Type::PATH, Expr::Type::VALUE, Expr::Type::SUBEXPR, Expr::Type::SUBSCRIPT))
-            return *err;
-
         auto _ = scopedNotInPath(p);
         auto body = p.parseTo(Token::RBRACE);
-        if (!body)
-            return unexpected<Error>(std::move(body.error()));
-
+        TRY_EXPECTED(body);
         return simplifyOrForward(p.env, std::make_unique<SubExpr>(std::move(left),
                                                                   std::move(*body)));
     }
@@ -637,7 +595,7 @@ public:
                 if (t.containsPoint(comp_->point)) {
                     return std::make_unique<CompletionWordExpr>(word.substr(0, comp_->point - t.begin), comp_, t);
                 }
-                return std::make_unique<ConstExpr>(Value::make<std::string>(std::move(word)));
+                return std::make_unique<CompletionConstExpr>(Value::make<std::string>(std::move(word)));
             }
             /* Constant */
             else if (auto constant = p.env->findConstant(word)) {
@@ -716,71 +674,114 @@ public:
     Completion* comp_;
 };
 
+namespace
+{
+// Static stateles parselets re-used by all parser instances
+const ScalarParser<int64_t> intParser;
+const ScalarParser<double> floatParser;
+const ScalarParser<std::string> stringParser;
+const RegExpParser regexpParser;
+const UnaryOpParser<OperatorNegate> negateParser;
+const UnaryOpParser<OperatorBitInv> bitInvParser;
+const UnaryOpParser<OperatorNot> notParser;
+const UnaryOpParser<OperatorLen> lenParser;
+const UnaryOpParser<OperatorTypeof> typeofParser;
+const UnaryPostOpParser<OperatorBool> boolParser;
+const BinaryOpParser<OperatorAdd, Precedence::TERM> addParser;
+const BinaryOpParser<OperatorSub, Precedence::TERM> subParser;
+const BinaryOpParser<OperatorMul, Precedence::PRODUCT> mulParser;
+const BinaryOpParser<OperatorDiv, Precedence::PRODUCT> divParser;
+const BinaryOpParser<OperatorMod, Precedence::PRODUCT> modParser;
+const BinaryOpParser<OperatorBitAnd, Precedence::BITWISE> bitAndParser;
+const BinaryOpParser<OperatorBitOr, Precedence::BITWISE> bitOrParser;
+const BinaryOpParser<OperatorBitXor, Precedence::BITWISE> bitXorParser;
+const BinaryOpParser<OperatorShl, Precedence::BITWISE> shlParser;
+const BinaryOpParser<OperatorShr, Precedence::BITWISE> shrParser;
+const BinaryOpParser<OperatorEq, Precedence::EQUALITY> eqParser;
+const BinaryOpParser<OperatorNeq, Precedence::EQUALITY> neqParser;
+const BinaryOpParser<OperatorLt, Precedence::EQUALITY> ltParser;
+const BinaryOpParser<OperatorLtEq, Precedence::EQUALITY> lteqParser;
+const BinaryOpParser<OperatorGt, Precedence::EQUALITY> gtParser;
+const BinaryOpParser<OperatorGtEq, Precedence::EQUALITY> gteqParser;
+const AndOrParser andOrParser;
+const CastParser castParser;
+const ParenParser parenParser;
+const SubSelectParser subSelectParser;
+const SubscriptParser subscriptParser;
+const WordParser wordParser;
+const PathParser pathParser;
+const UnpackOpParser unpackParser;
+const WordOpParser wordOpParser;
+const ConstParser trueParser{Value::t()};
+const ConstParser falseParser{Value::f()};
+const ConstParser nullParser{Value::null()};
+}
+
 static auto setupParser(Parser& p)
 {
     /* Scalars */
-    p.prefixParsers[Token::C_TRUE]  = std::make_unique<ConstParser>(Value::t());
-    p.prefixParsers[Token::C_FALSE] = std::make_unique<ConstParser>(Value::f());
-    p.prefixParsers[Token::C_NULL]  = std::make_unique<ConstParser>(Value::null());
-    p.prefixParsers[Token::INT]     = std::make_unique<ScalarParser<int64_t>>();
-    p.prefixParsers[Token::FLOAT]   = std::make_unique<ScalarParser<double>>();
-    p.prefixParsers[Token::STRING]  = std::make_unique<ScalarParser<std::string>>();
-    p.prefixParsers[Token::REGEXP]  = std::make_unique<RegExpParser>();
+    p.prefixParsers[Token::C_TRUE]  = &trueParser;
+    p.prefixParsers[Token::C_FALSE] = &falseParser;
+    p.prefixParsers[Token::C_NULL]  = &nullParser;
+    p.prefixParsers[Token::INT]     = &intParser;
+    p.prefixParsers[Token::FLOAT]   = &floatParser;
+    p.prefixParsers[Token::STRING]  = &stringParser;
+    p.prefixParsers[Token::REGEXP]  = &regexpParser;
 
     /* Unary Operators */
-    p.prefixParsers[Token::OP_SUB]    = std::make_unique<UnaryOpParser<OperatorNegate>>();
-    p.prefixParsers[Token::OP_BITINV] = std::make_unique<UnaryOpParser<OperatorBitInv>>();
-    p.prefixParsers[Token::OP_NOT]    = std::make_unique<UnaryOpParser<OperatorNot>>();
-    p.prefixParsers[Token::OP_LEN]    = std::make_unique<UnaryOpParser<OperatorLen>>();
-    p.infixParsers[Token::OP_BOOL]    = std::make_unique<UnaryPostOpParser<OperatorBool>>();
-    p.prefixParsers[Token::OP_TYPEOF] = std::make_unique<UnaryOpParser<OperatorTypeof>>();
-    p.infixParsers[Token::OP_UNPACK]  = std::make_unique<UnpackOpParser>();
-    p.infixParsers[Token::WORD]       = std::make_unique<WordOpParser>();
+    p.prefixParsers[Token::OP_SUB]    = &negateParser;
+    p.prefixParsers[Token::OP_BITINV] = &bitInvParser;
+    p.prefixParsers[Token::OP_NOT]    = &notParser;
+    p.prefixParsers[Token::OP_LEN]    = &lenParser;
+    p.infixParsers[Token::OP_BOOL]    = &boolParser;
+    p.prefixParsers[Token::OP_TYPEOF] = &typeofParser;
+    p.infixParsers[Token::OP_UNPACK]  = &unpackParser;
+    p.infixParsers[Token::WORD]       = &wordOpParser;
 
     /* Binary Operators */
-    p.infixParsers[Token::OP_ADD]   = std::make_unique<BinaryOpParser<OperatorAdd, Precedence::TERM>>();
-    p.infixParsers[Token::OP_SUB]   = std::make_unique<BinaryOpParser<OperatorSub, Precedence::TERM>>();
-    p.infixParsers[Token::OP_TIMES] = std::make_unique<BinaryOpParser<OperatorMul, Precedence::PRODUCT>>();
-    p.infixParsers[Token::OP_DIV]   = std::make_unique<BinaryOpParser<OperatorDiv, Precedence::PRODUCT>>();
-    p.infixParsers[Token::OP_MOD]   = std::make_unique<BinaryOpParser<OperatorMod, Precedence::PRODUCT>>();
+    p.infixParsers[Token::OP_ADD]   = &addParser;
+    p.infixParsers[Token::OP_SUB]   = &subParser;
+    p.infixParsers[Token::OP_TIMES] = &mulParser;
+    p.infixParsers[Token::OP_DIV]   = &divParser;
+    p.infixParsers[Token::OP_MOD]   = &modParser;
 
     /* Bit Operators */
-    p.infixParsers[Token::OP_BITAND] = std::make_unique<BinaryOpParser<OperatorBitAnd, Precedence::BITWISE>>();
-    p.infixParsers[Token::OP_BITOR]  = std::make_unique<BinaryOpParser<OperatorBitOr, Precedence::BITWISE>>();
-    p.infixParsers[Token::OP_BITXOR] = std::make_unique<BinaryOpParser<OperatorBitXor, Precedence::BITWISE>>();
-    p.infixParsers[Token::OP_LSHIFT] = std::make_unique<BinaryOpParser<OperatorShl, Precedence::BITWISE>>();
-    p.infixParsers[Token::OP_RSHIFT] = std::make_unique<BinaryOpParser<OperatorShr, Precedence::BITWISE>>();
+    p.infixParsers[Token::OP_BITAND] = &bitAndParser;
+    p.infixParsers[Token::OP_BITOR]  = &bitOrParser;
+    p.infixParsers[Token::OP_BITXOR] = &bitXorParser;
+    p.infixParsers[Token::OP_LSHIFT] = &shlParser;
+    p.infixParsers[Token::OP_RSHIFT] = &shrParser;
 
     /* Comparison/Test */
-    p.infixParsers[Token::OP_EQ]     = std::make_unique<BinaryOpParser<OperatorEq,   Precedence::EQUALITY>>();
-    p.infixParsers[Token::OP_NOT_EQ] = std::make_unique<BinaryOpParser<OperatorNeq,  Precedence::EQUALITY>>();
-    p.infixParsers[Token::OP_LT]     = std::make_unique<BinaryOpParser<OperatorLt,   Precedence::EQUALITY>>();
-    p.infixParsers[Token::OP_LTEQ]   = std::make_unique<BinaryOpParser<OperatorLtEq, Precedence::EQUALITY>>();
-    p.infixParsers[Token::OP_GT]     = std::make_unique<BinaryOpParser<OperatorGt,   Precedence::EQUALITY>>();
-    p.infixParsers[Token::OP_GTEQ]   = std::make_unique<BinaryOpParser<OperatorGtEq, Precedence::EQUALITY>>();
-    p.infixParsers[Token::OP_AND]    = std::make_unique<AndOrParser>();
-    p.infixParsers[Token::OP_OR]     = std::make_unique<AndOrParser>();
+    p.infixParsers[Token::OP_EQ]     = &eqParser;
+    p.infixParsers[Token::OP_NOT_EQ] = &neqParser;
+    p.infixParsers[Token::OP_LT]     = &ltParser;
+    p.infixParsers[Token::OP_LTEQ]   = &lteqParser;
+    p.infixParsers[Token::OP_GT]     = &gtParser;
+    p.infixParsers[Token::OP_GTEQ]   = &gteqParser;
+    p.infixParsers[Token::OP_AND]    = &andOrParser;
+    p.infixParsers[Token::OP_OR]     = &andOrParser;
 
     /* Cast */
-    p.infixParsers[Token::OP_CAST]   = std::make_unique<CastParser>();
+    p.infixParsers[Token::OP_CAST]   = &castParser;
 
     /* Subexpressions/Subscript */
-    p.prefixParsers[Token::LPAREN] = std::make_unique<ParenParser>();     /* (...) */
-    p.prefixParsers[Token::LBRACE] = std::make_unique<SubSelectParser>(); /* {...} */
-    p.infixParsers[Token::LBRACE] = std::make_unique<SubSelectParser>();
-    p.prefixParsers[Token::LBRACK] = std::make_unique<SubscriptParser>(); /* [...] */
-    p.infixParsers[Token::LBRACK] = std::make_unique<SubscriptParser>();
+    p.prefixParsers[Token::LPAREN] = &parenParser;     /* (...) */
+    p.prefixParsers[Token::LBRACE] = &subSelectParser; /* {...} */
+    p.infixParsers[Token::LBRACE] = &subSelectParser;
+    p.prefixParsers[Token::LBRACK] = &subscriptParser; /* [...] */
+    p.infixParsers[Token::LBRACK] = &subscriptParser;
 
     /* Ident/Function */
-    p.prefixParsers[Token::WORD] = std::make_unique<WordParser>();
-    p.prefixParsers[Token::SELF] = std::make_unique<WordParser>();
+    p.prefixParsers[Token::WORD] = &wordParser;
+    p.prefixParsers[Token::SELF] = &wordParser;
 
     /* Wildcards */
-    p.prefixParsers[Token::WILDCARD] = std::make_unique<WordParser>();
-    p.prefixParsers[Token::OP_TIMES] = std::make_unique<WordParser>();
+    p.prefixParsers[Token::WILDCARD] = &wordParser;
+    p.prefixParsers[Token::OP_TIMES] = &wordParser;
 
     /* Paths */
-    p.infixParsers[Token::DOT]  = std::make_unique<PathParser>();
+    p.infixParsers[Token::DOT]  = &pathParser;
 }
 
 auto compile(Environment& env, std::string_view query, bool any, bool autoWildcard) -> expected<ASTPtr, Error>
@@ -832,19 +833,36 @@ auto complete(Environment& env, std::string_view query, size_t point, const Mode
     if (options.limit > 0)
         comp.limit = options.limit;
 
-    p.prefixParsers[Token::WORD]  = std::make_unique<CompletionWordParser>(&comp);
-    p.infixParsers[Token::DOT]    = std::make_unique<CompletionPathParser>(&comp);
-    p.infixParsers[Token::OP_AND] = std::make_unique<CompletionAndOrParser>(&comp);
-    p.infixParsers[Token::OP_OR]  = std::make_unique<CompletionAndOrParser>(&comp);
+    CompletionWordParser wordCompletionParser(&comp);
+    CompletionPathParser pathCompletionParser(&comp);
+    CompletionAndOrParser andOrCompletionParser(&comp);
+    p.prefixParsers[Token::WORD]  = &wordCompletionParser;
+    p.infixParsers[Token::DOT]    = &pathCompletionParser;
+    p.infixParsers[Token::OP_AND] = &andOrCompletionParser;
+    p.infixParsers[Token::OP_OR]  = &andOrCompletionParser;
 
     auto astResult = p.parse();
+    if (!p.match(Token::Type::NIL))
+        return unexpected<Error>(Error::ExpectedEOF, "Expected end-of-input; got "s + p.current().toString());
+
     TRY_EXPECTED(astResult);
     auto ast = std::move(*astResult);
 
-    /* Expand a single value to `** == <value>` */
-    if (options.autoWildcard && ast && ast->constant()) {
-        ast = std::make_unique<BinaryExpr<OperatorEq>>(
-            std::make_unique<WildcardExpr>(), std::move(ast));
+    // Determine which hints to show.
+    auto showConstantWildcardHint = false;
+    auto showFieldWildcardHint = false;
+    auto showComparisonWildcardHint = false;
+    if (options.showWildcardHints) {
+        // Test the query for patterns and hint for converting it
+        // to a wildcard query by prepending `**.` to the query.
+        if (isSingleValueExpression(ast.get()))
+            showConstantWildcardHint = true;
+
+        if (isSingleValueOrFieldExpression(ast.get()))
+            showFieldWildcardHint = true;
+
+        if (isFieldComparison(ast.get()))
+            showComparisonWildcardHint = true;
     }
 
     Context ctx(&env);
@@ -861,6 +879,25 @@ auto complete(Environment& env, std::string_view query, size_t point, const Mode
             return left.text < right.text;
         });
 
+    // Show special hints for wildcard expansion.
+    if (showFieldWildcardHint)
+        candidates.emplace_back(fmt::format("**.{}", query),
+                                SourceLocation(0, query.size()),
+                                CompletionCandidate::Type::HINT,
+                                fmt::format("Query field '{}' recursive", query));
+
+    if (showConstantWildcardHint)
+        candidates.emplace_back(fmt::format("** = {}", query),
+                                SourceLocation(0, query.size()),
+                                CompletionCandidate::Type::HINT,
+                                fmt::format("Query fields matching '{}' recursive", query));
+
+    if (showComparisonWildcardHint)
+        candidates.emplace_back(fmt::format("**.{}", query),
+                                SourceLocation(0, query.size()),
+                                CompletionCandidate::Type::HINT,
+                                "Expand to recursive query");
+
     return candidates;
 }
 
@@ -873,17 +910,18 @@ auto eval(Environment& env, const AST& ast, const ModelNode& node, Diagnostics* 
 
     auto mutableAST = ast.expr().clone();
 
-    std::vector<Value> res;
-    mutableAST->eval(ctx, Value::field(node), LambdaResultFn([&res](Context, Value value) {
-        res.push_back(std::move(value));
+    std::vector<Value> values;
+    auto res = mutableAST->eval(ctx, Value::field(node), LambdaResultFn([&values](Context, Value&& value) {
+        values.push_back(std::move(value));
         return Result::Continue;
     }));
+    TRY_EXPECTED(res);
 
     if (diag) {
         diag->collect(*mutableAST);
     }
 
-    return res;
+    return values;
 }
 
 auto diagnostics(Environment& env, const AST& ast, const Diagnostics& diag) -> expected<std::vector<Diagnostics::Message>, Error>
