@@ -119,7 +119,9 @@ static auto result(const ModelPoolPtr& model, std::string_view query)
     REQUIRE(ast.has_value());
     INFO("AST: " << (*ast)->expr().toString());
 
-    return eval(env, **ast, *model->root(0), nullptr);
+    auto root = model->root(0);
+    REQUIRE(root);
+    return eval(env, **ast, **root, nullptr);
 }
 
 static auto joined_result(const ModelPoolPtr& model, std::string_view query)
@@ -153,6 +155,10 @@ TEST_CASE("Big model query performance", "[perf.big-model-benchmark]") {
     BENCHMARK("Query field id") {
         return result(model, "count(*.id == 25)");
     };
+
+    BENCHMARK("Query field keys recursive") {
+        return result(model, "count(keys(**))");
+    };
 }
 
 #define REQUIRE_RESULT(query, result) \
@@ -166,4 +172,19 @@ TEST_CASE("Big model queries", "[perf.big-model-queries]") {
     REQUIRE_RESULT("count(**.id == 250)", "1");
     REQUIRE_RESULT("count(*.id == 250)", "1");
     CALLGRIND_STOP_INSTRUMENTATION;
+}
+
+TEST_CASE("Slow Compile Time Query", "[perf.slow-compile-time-query]") {
+    if (RUNNING_ON_VALGRIND) {
+        SKIP("Skipping benchmarks when running under valgrind");
+    }
+
+    const auto model = generate_model(1);
+
+    BENCHMARK("Find Primes") {
+        auto value = result(model, "count(range(1,1000)...{count((_ % range(1,_)...) == 0) == 2})");
+        REQUIRE(value);
+        REQUIRE(value->front().template as<ValueType::Int>() == 168);
+        return value;
+    };
 }
